Refactor wine-vanilla-pba, wine-staging-pba patchsets
Refactor the per-commit wine-pba ordered patchsets so patches are
per now per-file. Patches are easier to rebase and can be applied
out-of-order. Hard links are used to reduce the archive size.

Add a small hack to better support Wine (Staging) 3.17 and 3.18.
However Wine (Staging) 3.19(+) remains broken. Due to use of ARB
buffer objects in stock Wine.